=fixing problem with lazyAction not selecting the correct export from modules that has more then a single default export. Quick fix is to move createPlugin default export into a standalone file.
